Make sure all data manipulation around mimeinfo->length and mimeinfo->offset use at least long integers (they are more or less all bound to ftell), this solves mismatch in comparisons and also: CID 1498802: Improper use of negative value (NEGATIVE_RETURNS)

+++ src/mimeview.c
@@ -181,7 +181,7 @@ static void mimeview_copy_cb(GtkAction *action, gpointer data)
 	} else {
 		void *data = procmime_get_part_as_string(mimeinfo, FALSE);
 		gtk_clipboard_set_text(gtk_clipboard_get(GDK_SELECTION_CLIPBOARD),
-				       data, mimeinfo->length);
+				       data, (gint)mimeinfo->length);
 		g_free(data);
 	}
 }
